# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cottage Village API Documentation",
        "description": "",
        "contact": {
            "email": "afonico@info.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register user",
                "description": "",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "birthday",
                                    "gender_id",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Users name. Size should be less than 100 characters",
                                        "type": "string"
                                    },
                                    "surname": {
                                        "description": "Users surname. Size should be less than 100 characters",
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "description": "Users birthday date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender_id": {
                                        "description": "User gender_id, 1 = Male, 2 = Female",
                                        "type": "integer",
                                        "default": 1,
                                        "enum": [
                                            1,
                                            2
                                        ]
                                    },
                                    "phone": {
                                        "description": "Users phone. Unique value, should be masked value in format +7 (111) 222-33-44, server proceed to format +71112223344 to save in DB and returns to client in masked format",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Users email. Unique value",
                                        "type": "string"
                                    },
                                    "whatsapp_phone": {
                                        "description": "Should be masked value in format +7 (111) 222-33-44",
                                        "type": "string"
                                    },
                                    "telegram_phone": {
                                        "description": "Should be masked value in format +7 (111) 222-33-44",
                                        "type": "string"
                                    },
                                    "viber_phone": {
                                        "description": "Should be masked value in format +7 (111) 222-33-44",
                                        "type": "string"
                                    },
                                    "work_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "hobby_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "family_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "extra_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "user_estates": {
                                        "description": "User estates data. estate_type_id variations is { 1: 'Участок', 2: 'Квартира'}. estate_number should be greater or equal 0. estate_number required when picked estate_type_id",
                                        "type": "string",
                                        "format": "json",
                                        "example": "{ 'user_estates': [ { 'estate_type_id': 1, estate_number: 10 } ] }"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/password/reset/send": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Reset user password by email. Token from link expires in 1 hour",
                "description": "",
                "operationId": "passwordResetNotification",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Users email. Unique value",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "If the user is found, the email will be sent"
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Reset user password by email",
                "description": "",
                "operationId": "passwordReset",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Users new password. Minimum 8 symbols",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Users new password cpnfirmation. Minimum 8 symbols",
                                        "type": "string"
                                    },
                                    "token": {
                                        "description": "Users verification token from email. Token expires in 1 hour",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: token_type access_token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}